Refactor split and analize into modules
