+++ go.mod
@@ -1,0 +1,3 @@
+module gohz
+
+go 1.24.3
